requirements.txt
tornado==4.4.2
bcrypt==3.1.1
pyjwt==1.4.0
mongoengine==0.10.6
websocket-client==0.37.0
celery==3.1.23
requests==2.9.1
ldap3==1.4.0